chore: move processDepositRequest() to onboardBuilder.ts
